Fast test now checks borders
Added border checking to the fast method so as it steps out of its calls
it looks for possible closer points on the opposite side of each
boundary around it.

Accuracy seems to have increased.

Some improper calls are still being made about 3/10 of the time. (though
the bad calls are still very close)

additions:
-accuracy testing code
-border checking in fast solution