this commit is broken! but it's at 100% now
Cleaned up the search code so it's not a terrifying cludge of if-else's
(still looks pretty bad, but I don't know what I can do to make it
better without losing accuracy)

Accuracy is at 100% all the time. Not even using that 25% error margin.

Speed is bad.... so I'm working on sorting the list nicely by median
values....
That's why it's broken