Turn-in set
The program is 100% accurate.

The speed of the program is significantly less than the speed of my
linear comparison, however I think that the worst case scenario of my
speed has linear growth. The speed on 100,000 function calls varied by
an order of magnitude, so the randomness of how my list is shuffled
makes a big difference to how efficient my tree is (height is too big).

I attempted to pre-sort the tree in my build method and a special insert
method that should have organized by the median values. However, I was
saddened to find that the running time seemed to be higher than my
randomized list insert. So those methods are commented out.